fix: redirect HuggingFace model cache to writable directory
@huggingface/transformers defaults to caching models inside its own
package directory in node_modules (.cache/), which is read-only on
macOS and in some Docker setups. This caused Kokoro TTS model downloads
to fail during initial setup.

Extract a shared ensureModelCacheDir() utility that sets env.cacheDir
on the @huggingface/transformers runtime config (the actual mechanism
the JS library uses) and creates the directory if needed. Apply it to
all three consumers: TTS (kokoro-js), STT (whisper), and embeddings.

Closes #357
